Project/requirements.txt
blinker<=1.6.3
certifi<=2023.11.17
charset-normalizer<=3.3.2
click<=8.1.7
filelock<=3.12.2
Flask<=3.0.0
fsspec<=2023.12.0
huggingface-hub<=0.19.4
idna<=3.6
itsdangerous<=2.1.2
Jinja2<=3.1.2
MarkupSafe<=2.1.3
mpmath<=1.3.0
networkx<=3.2.1
numpy<=1.26.2
packaging<=23.2
Pillow<=10.1.0
PyYAML<=6.0.1
regex<=2023.10.3
requests<=2.31.0
safetensors<=0.4.1
sympy<=1.12
tokenizers<0.14.0
torch<=2.1.1
torchvision<=0.16.1
tqdm<=4.66.1
transformers<=4.35.2
typing_extensions<=4.8.0
urllib3<=2.1.0
Werkzeug<=3.0.1
